pilot/requirements.txt
cachetools >=2.1, <3.0
jsoncomment >=0.3, <1.0
pid-controller >=0.2, <1.0
pyzmq >=18.1, <19.0
